backend/requirements.txt
aniso8601==9.0.1
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
distlib==0.3.4
Django==4.0.2
docopt==0.6.2
fastapi==0.70.0
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
image==1.5.33
intel-openmp==2022.0.2
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
numpy==1.22.2
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.5.1
pyasn1==0.4.8
pydantic==1.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
starlette==0.16.0
sudo==1.0.0
toml==0.10.2
torch==1.10.2+cpu
torchvision==0.11.3+cpu
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
uvicorn==0.15.0
virtualenv==20.13.2
virtualenv-clone==0.5.7
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
yarg==0.1.9
-f https://download.pytorch.org/whl/torch_stable.html

